dashboardBackend/requirements.txt
django
selenium
pandas
redis
redislite
pyarrow
djangorestframework
datetime
django-cors-headers
apscheduler
zipfile36